# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shotgun Ninjas Video Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StatsOverview {
  totalProjects: number;
  totalScenes: number;
  totalPrompts: number;
  totalExports: number;
  analyzedProjects: number;
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  project_created: "project_created",
  audio_uploaded: "audio_uploaded",
  analyzed: "analyzed",
  storyboard_generated: "storyboard_generated",
  prompts_generated: "prompts_generated",
  exported: "exported",
} as const;

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  message: string;
  projectId: string;
  projectTitle?: string;
  createdAt: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  draft: "draft",
  uploaded: "uploaded",
  analyzed: "analyzed",
  storyboarded: "storyboarded",
  prompted: "prompted",
  exported: "exported",
} as const;

export interface Project {
  id: string;
  title: string;
  artist?: string;
  genre?: string;
  mood?: string;
  visualDirection?: string;
  status: ProjectStatus;
  coverColor?: string;
  createdAt: string;
  updatedAt: string;
  sceneCount?: number;
  promptCount?: number;
  exportCount?: number;
  durationSec?: number;
  bpm?: number;
  keySignature?: string;
}

export interface AudioFile {
  id: string;
  projectId: string;
  fileName: string;
  mimeType: string;
  sizeBytes: number;
  durationSec?: number;
  createdAt: string;
}

export type TimelineSegmentSection =
  (typeof TimelineSegmentSection)[keyof typeof TimelineSegmentSection];

export const TimelineSegmentSection = {
  intro: "intro",
  verse: "verse",
  pre_chorus: "pre_chorus",
  chorus: "chorus",
  bridge: "bridge",
  drop: "drop",
  breakdown: "breakdown",
  outro: "outro",
} as const;

export interface TimelineSegment {
  id: string;
  projectId: string;
  index: number;
  startSec: number;
  endSec: number;
  section: TimelineSegmentSection;
  /** 0..1 */
  intensity: number;
  emotion: string;
  bpm?: number;
}

export interface EmotionalPoint {
  timeSec: number;
  valence: number;
  arousal: number;
  label: string;
}

export interface AnalysisResult {
  projectId: string;
  durationSec: number;
  bpm: number;
  keySignature: string;
  energy: number;
  loudnessDb?: number;
  segments: TimelineSegment[];
  emotionalMap: EmotionalPoint[];
}

export type ProjectDetail = Project & {
  audio?: AudioFile;
  analysis?: AnalysisResult;
};

export interface CreateProjectInput {
  /** @minLength 1 */
  title: string;
  artist?: string;
  genre?: string;
  mood?: string;
  visualDirection?: string;
}

export type UpdateProjectInputStatus =
  (typeof UpdateProjectInputStatus)[keyof typeof UpdateProjectInputStatus];

export const UpdateProjectInputStatus = {
  draft: "draft",
  uploaded: "uploaded",
  analyzed: "analyzed",
  storyboarded: "storyboarded",
  prompted: "prompted",
  exported: "exported",
} as const;

export interface UpdateProjectInput {
  title?: string;
  artist?: string;
  genre?: string;
  mood?: string;
  visualDirection?: string;
  status?: UpdateProjectInputStatus;
}

export interface UploadAudioInput {
  fileName: string;
  sizeBytes: number;
  mimeType: string;
  durationSec?: number;
}

export interface StoryboardScene {
  id: string;
  projectId: string;
  segmentId?: string;
  index: number;
  startSec: number;
  endSec: number;
  title: string;
  description: string;
  shotType: string;
  cameraMovement: string;
  location: string;
  lighting: string;
  colorPalette: string;
  wardrobe?: string;
  notes?: string;
}

export interface UpdateSceneInput {
  title?: string;
  description?: string;
  shotType?: string;
  cameraMovement?: string;
  location?: string;
  lighting?: string;
  colorPalette?: string;
  wardrobe?: string;
  notes?: string;
}

export type PromptModel = (typeof PromptModel)[keyof typeof PromptModel];

export const PromptModel = {
  runway: "runway",
  sora: "sora",
  kling: "kling",
  pika: "pika",
  luma: "luma",
  generic: "generic",
} as const;

export interface Prompt {
  id: string;
  projectId: string;
  sceneId: string;
  index: number;
  model: PromptModel;
  text: string;
  negativePrompt?: string;
  aspectRatio?: string;
  durationSec?: number;
}

export type UpdatePromptInputModel =
  (typeof UpdatePromptInputModel)[keyof typeof UpdatePromptInputModel];

export const UpdatePromptInputModel = {
  runway: "runway",
  sora: "sora",
  kling: "kling",
  pika: "pika",
  luma: "luma",
  generic: "generic",
} as const;

export interface UpdatePromptInput {
  text?: string;
  negativePrompt?: string;
  model?: UpdatePromptInputModel;
  aspectRatio?: string;
  durationSec?: number;
}

export type ExportRecordFormat =
  (typeof ExportRecordFormat)[keyof typeof ExportRecordFormat];

export const ExportRecordFormat = {
  json: "json",
  txt: "txt",
  production_plan: "production_plan",
} as const;

export interface ExportRecord {
  id: string;
  projectId: string;
  format: ExportRecordFormat;
  content: string;
  createdAt: string;
}

export type CreateExportInputFormat =
  (typeof CreateExportInputFormat)[keyof typeof CreateExportInputFormat];

export const CreateExportInputFormat = {
  json: "json",
  txt: "txt",
  production_plan: "production_plan",
} as const;

export interface CreateExportInput {
  format: CreateExportInputFormat;
}

export type SettingsDefaultModel =
  (typeof SettingsDefaultModel)[keyof typeof SettingsDefaultModel];

export const SettingsDefaultModel = {
  runway: "runway",
  sora: "sora",
  kling: "kling",
  pika: "pika",
  luma: "luma",
  generic: "generic",
} as const;

export type SettingsTheme = (typeof SettingsTheme)[keyof typeof SettingsTheme];

export const SettingsTheme = {
  dark: "dark",
  light: "light",
} as const;

export interface Settings {
  defaultModel: SettingsDefaultModel;
  defaultAspectRatio: string;
  defaultSceneDurationSec: number;
  theme: SettingsTheme;
  creatorName?: string;
  creatorHandle?: string;
}

export type UpdateSettingsInputDefaultModel =
  (typeof UpdateSettingsInputDefaultModel)[keyof typeof UpdateSettingsInputDefaultModel];

export const UpdateSettingsInputDefaultModel = {
  runway: "runway",
  sora: "sora",
  kling: "kling",
  pika: "pika",
  luma: "luma",
  generic: "generic",
} as const;

export type UpdateSettingsInputTheme =
  (typeof UpdateSettingsInputTheme)[keyof typeof UpdateSettingsInputTheme];

export const UpdateSettingsInputTheme = {
  dark: "dark",
  light: "light",
} as const;

export interface UpdateSettingsInput {
  defaultModel?: UpdateSettingsInputDefaultModel;
  defaultAspectRatio?: string;
  defaultSceneDurationSec?: number;
  theme?: UpdateSettingsInputTheme;
  creatorName?: string;
  creatorHandle?: string;
}