# lib/api-zod/src/generated/types/audioFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shotgun Ninjas Video Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface AudioFile {
  id: string;
  projectId: string;
  fileName: string;
  mimeType: string;
  sizeBytes: number;
  durationSec?: number;
  createdAt: Date;
}